requirements.txt
flask
flask-cors
pypdf
python-dotenv
pathlib
gunicorn
google-genai